More progress on CompCert 3.1 port

+++ veric/SeparationLogic.v
@@ -70,7 +70,7 @@ Transparent mpred Nveric Sveric Cveric Iveric Rveric Sveric SIveric CSLveric CIv
 Definition denote_tc_iszero v : mpred :=
          match v with
          | Vint i => prop (is_true (Int.eq i Int.zero))
-         | Vlong i => prop (is_true (Int.eq (Int.repr (Int64.unsigned i)) Int.zero))
+         | Vlong i => prop (is_true (Int64.eq (Int64.repr (Int64.unsigned i)) Int64.zero))
          | _ => FF
          end.
